Move partialeq to base impl
